refactor: 대기 취소 시 검증 한꺼번에 진행하도록 수정
